+++ src/pages/PROverview/PullRequestOverview.tsx
@@ -2,7 +2,10 @@ import React from "react";
 import { Box, Text } from "ink";
 import { MAIN_COLOR } from "../../theme/colors.js";
 import { REVIEW_HEADLINE_TEXT } from "../../theme/banners.js";
-import { PullRequestDetails, SpecReview } from "../../lib/clients/baz.js";
+import type {
+  PullRequestDetails,
+  SpecReview,
+} from "../../lib/providers/index.js";
 import { Issue } from "../../issues/types.js";
 
 const CHECKBOX_PLACEHOLDER = " □ ";
