+++ src/pages/PRSelector/PullRequestSelectorContainer.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Box, Text, useInput } from "ink";
 import Spinner from "ink-spinner";
-import { PullRequest } from "../../lib/clients/baz.js";
+import type { PullRequest } from "../../lib/providers/index.js";
 import { usePullRequests } from "../../hooks/usePullRequests.js";
 import PullRequestSelector from "./PullRequestSelector.js";
